Fix clickless login in CI
