+++ frontend/src/e2e-test/config/users.ts
@@ -5,11 +5,24 @@
 // This will be used with Testcafe's useRole.
 
 import { Role } from 'testcafe'
-import { DevLoginUser } from '../pages/dev-login-form'
 import Home from '../pages/home'
 import config from 'e2e-test-common/config'
+import { UUID } from 'e2e-test-common/dev-api/types'
 
 const home = new Home()
+
+export interface DevLoginUser {
+  aad: UUID
+  roles: DevLoginRole[]
+}
+
+export type DevLoginRole = typeof devLoginRoles[number]
+const devLoginRoles = [
+  'SERVICE_WORKER',
+  'FINANCE_ADMIN',
+  'DIRECTOR',
+  'ADMIN'
+] as const
 
 export const seppoAdmin: DevLoginUser = {
   aad: config.adminAad,
@@ -26,9 +39,8 @@ export async function employeeLogin(
   user: string | DevLoginUser,
   landingUrl?: string
 ) {
-  await t.navigateTo(config.employeeUrl)
-  const authUrl =
-    'http://localhost:9099/api/internal/auth/saml/login/callback?RelayState=%2Femployee'
+  await t.navigateTo(config.employeeLoginUrl)
+  const authUrl = `${config.apiUrl}/auth/saml/login/callback?RelayState=%2Femployee`
 
   await t.eval(
     () => {
@@ -43,9 +55,17 @@ export async function employeeLogin(
         params.append('aad', user.aad)
         user.roles.forEach((role) => params.append('roles', role))
       }
-      return fetch(authUrl, { method: 'POST', body: params }).then(
-        () => undefined
-      )
+      return fetch(authUrl, {
+        method: 'POST',
+        body: params,
+        redirect: 'manual'
+      }).then((response) => {
+        if (response.status >= 400) {
+          throw new Error(
+            `Fetch to {authUrl} failed with status ${response.status}`
+          )
+        }
+      })
     },
     { dependencies: { authUrl, user } }
   )
@@ -63,7 +83,7 @@ export async function mobileLogin(t: TestController, token: string) {
 export const enduserRole = Role(
   config.enduserUrl,
   async () => {
-    await home.login('enduser')
+    await home.enduserLogin()
   },
   { preserveUrl: true }
 )
